@@ -1,27 +1,37 @@
 {
-  "main": "node_modules/expo/AppEntry.js",
+  "name": "Pata",
+  "version": "0.0.1",
+  "private": true,
   "scripts": {
-    "start": "expo start",
-    "android": "expo start --android",
-    "ios": "expo start --ios",
-    "web": "expo start --web"
+    "android": "react-native run-android",
+    "ios": "react-native run-ios",
+    "start": "react-native start",
+    "test": "jest",
+    "lint": "eslint ."
   },
   "dependencies": {
-    "expo": "~46.0.17",
-    "expo-status-bar": "~1.4.0",
-    "react": "18.0.0",
-    "react-native": "0.69.6",
-    "react-dom": "18.0.0",
-    "react-native-screens": "~3.15.0",
-    "@react-navigation/stack": "*",
-    "@react-navigation/native": "6.0.10",
-    "react-native-gesture-handler": "~2.5.0",
-    "@react-navigation/native-stack": "6.6.1",
-    "react-native-safe-area-context": "4.3.1",
-    "@testing-library/jest-dom": "*"
+    "@react-navigation/native": "^6.1.0",
+    "@react-navigation/native-stack": "^6.9.5",
+    "crypto-js": "^4.1.1",
+    "react": "18.1.0",
+    "react-native": "0.70.6",
+    "react-native-fs": "^2.20.0",
+    "react-native-safe-area-context": "^4.4.1",
+    "react-native-screens": "^3.18.2",
+    "react-native-sqlite-storage": "^6.0.1",
+    "react-navigation": "^4.4.4"
   },
   "devDependencies": {
-    "@babel/core": "^7.12.9"
+    "@babel/core": "^7.12.9",
+    "@babel/runtime": "^7.12.5",
+    "@react-native-community/eslint-config": "^2.0.0",
+    "babel-jest": "^26.6.3",
+    "eslint": "^7.32.0",
+    "jest": "^26.6.3",
+    "metro-react-native-babel-preset": "0.72.3",
+    "react-test-renderer": "18.1.0"
   },
-  "private": true
+  "jest": {
+    "preset": "react-native"
+  }
 }
