@@ -1,37 +1,36 @@
-const { InvalidArgumentError } = require('../../errors/invalidArgumentError')
+import {InvalidArgumentError} from '../../errors/invalidArgumentError';
 
-const { ESizeType } = require("../../enums/sizeType")
+import {ESizeType} from '../../enums/sizeType';
 
-class AnimaValueObject {
-    race
-    size
+export default class AnimaValueObject {
+  race;
+  size;
 
-    constructor(race, size) {
-        this.race = race
-        this.size = size
-        this.validateValueObject()
+  constructor(race, size) {
+    this.race = race;
+    this.size = size;
+    this.validateValueObject();
+  }
+
+  validateValueObject() {
+    this.validateRace(this.race);
+    this.validateSize(this.size);
+  }
+
+  validateRace(race) {
+    if (!race || race === '') {
+      throw new InvalidArgumentError('race argument cannot be null or empty');
     }
+  }
 
-    validateValueObject() {
-        this.validateRace(this.race)
-        this.validateSize(this.size)
+  validateSize(size) {
+    const values = Object.values(ESizeType);
+    if (!size || !values.includes(size)) {
+      throw new InvalidArgumentError(
+        `size argument cannot be null and must have one of the following values: ${values.join(
+          ', ',
+        )}`,
+      );
     }
-
-    validateRace(race) {
-        if (!race || race === "") {
-            throw new InvalidArgumentError("race argument cannot be null or empty")
-        }
-    }
-
-    validateSize(size) {
-        const values = Object.values(ESizeType)
-        if (!size || !values.includes(size)) {
-            throw new InvalidArgumentError(
-                `size argument cannot be null and must have one of the following values: ${values.join(', ')}`)
-        }
-    }
-}
-
-module.exports = {
-    AnimaValueObject
+  }
 }
