@@ -1,24 +1,22 @@
-const { InvalidArgumentError } = require("../../errors/invalidArgumentError")
+import {InvalidArgumentError} from '../../errors/invalidArgumentError';
 
-class ImageValueObject {
-    base64Image
+export default class ImageValueObject {
+  base64Image;
 
-    constructor(base64Image) {
-        this.base64Image = base64Image
-        this.validateValueObject()
+  constructor(base64Image) {
+    this.base64Image = base64Image;
+    this.validateValueObject();
+  }
+
+  validateValueObject() {
+    this.validateBase64Image(this.base64Image);
+  }
+
+  validateBase64Image(base64Image) {
+    if (!base64Image || base64Image === '') {
+      throw new InvalidArgumentError(
+        'base64Iamge argument cannot be null and empty',
+      );
     }
-
-    validateValueObject() {
-        this.validateBase64Image(this.base64Image)
-    }
-
-    validateBase64Image(base64Image) {
-        if (!base64Image || base64Image === "") {
-            throw new InvalidArgumentError("base64Iamge argument cannot be null and empty")
-        }
-    }
-}
-
-module.exports = {
-    ImageValueObject
+  }
 }
