@@ -1,34 +1,30 @@
-const { InvalidArgumentError } = require("../../errors/invalidArgumentError")
+import {InvalidArgumentError} from '../../errors/invalidArgumentError';
 
-class LocalizationValueObject {
-    city
-    state
+export default class LocalizationValueObject {
+  city;
+  state;
 
-    constructor(city, state) {
-        this.city = city
-        this.state = state
+  constructor(city, state) {
+    this.city = city;
+    this.state = state;
 
-        this.validateValueObject()
+    this.validateValueObject();
+  }
+
+  validateValueObject() {
+    this.validateCity(this.city);
+    this.validateState(this.state);
+  }
+
+  validateCity(city) {
+    if (!city || city === '') {
+      throw new InvalidArgumentError('city argument cannot be null and empty');
     }
+  }
 
-    validateValueObject() {
-        this.validateCity(this.city)
-        this.validateState(this.state)
+  validateState(state) {
+    if (!state || state === '') {
+      throw new InvalidArgumentError('state argument cannot be null and empty');
     }
-
-    validateCity(city) {
-        if (!city || city === "") {
-            throw new InvalidArgumentError("city argument cannot be null and empty")
-        }
-    }
-
-    validateState(state) {
-        if (!state || state === "") {
-            throw new InvalidArgumentError("state argument cannot be null and empty")
-        }
-    }
-}
-
-module.exports = {
-    LocalizationValueObject
+  }
 }
