@@ -1,34 +1,31 @@
-const { InvalidArgumentError } = require("../../errors/invalidArgumentError")
+import InvalidArgumentError from '../../errors/invalidArgumentError';
 
-class NameValueObject {
-    firstName
-    lastName
+export default class NameValueObject {
+  constructor(firstName, lastName) {
+    this.firstName = firstName;
+    this.lastName = lastName;
 
-    constructor(firstName, lastName) {
-        this.firstName = firstName
-        this.lastName = lastName
+    this.validateValueObject();
+  }
 
-        this.validateValueObject()
+  validateValueObject() {
+    this.validateFirstName(this.firstName);
+    this.validateLastName(this.lastName);
+  }
+
+  validateFirstName(firstName) {
+    if (!firstName || firstName === '') {
+      throw new InvalidArgumentError(
+        'firstName argument cannot be null and empty',
+      );
     }
+  }
 
-    validateValueObject() {
-        this.validateFirstName(this.firstName)
-        this.validateLastName(this.lastName)
-    }
-
-    validateFirstName(firstName) {
-        if (!firstName || firstName === "") {
-            throw new InvalidArgumentError("firstName argument cannot be null and empty")
-        }
-    }
-
-    validateLastName(lastName) {
-        if (!lastName || lastName === "") {
-            throw new InvalidArgumentError("lastName argument cannot be null and empty")
-        }
-    }
-}
-
-module.exports = {
-    NameValueObject
+  validateLastName(lastName) {
+    // if (!lastName || lastName === '') {
+    //   throw new InvalidArgumentError(
+    //     'lastName argument cannot be null and empty',
+    //   );
+    // }
+  }
 }
