@@ -1,9 +1,5 @@
-const ESizeType = {
-    smallSize: "smallSize",
-    midsize: "midSize",
-    large: "large"
-}
-
-module.exports = {
-    ESizeType
-}
+export const ESizeType = {
+  smallSize: 'smallSize',
+  midsize: 'midSize',
+  large: 'large',
+};
