@@ -1,10 +1,6 @@
-class InvalidArgumentError extends Error {
-    constructor(message) {
-        super(message)
-        this.name = this.constructor.name
-    }
-}
-
-module.exports = {
-    InvalidArgumentError
+export default class InvalidArgumentError extends Error {
+  constructor(message) {
+    super(message);
+    this.name = this.constructor.name;
+  }
 }
