@@ -1,10 +1,6 @@
-class RegisteredEmailError extends Error {
-    constructor(message) {
-        super(message)
-        this.name = this.constructor.name
-    }
-}
-
-module.exports = {
-    RegisteredEmailError
+export default class RegisteredEmailError extends Error {
+  constructor(message) {
+    super(message);
+    this.name = this.constructor.name;
+  }
 }
