@@ -1,44 +1,44 @@
-const { UserEntity } = require("../aggregate/userEntity")
-const { ContactValueObject } = require("../aggregate/valueObject/contactValueObject")
-const { NameValueObject } = require("../aggregate/valueObject/nameValueObject")
-const { RegisteredEmailError } = require("../errors/registeredEmailError")
+import UserEntity from '../aggregate/userEntity';
+import ContactValueObject from '../aggregate/valueObject/contactValueObject';
+import NameValueObject from '../aggregate/valueObject/nameValueObject';
+import RegisteredEmailError from '../errors/registeredEmailError';
 
-class RegistrationService {
-    constructor(userService, encryptionService) {
-        this.userService = userService
-        this.encryptionService = encryptionService
+export default class RegistrationService {
+  constructor(userService, encryptionService) {
+    this.userService = userService;
+    this.encryptionService = encryptionService;
+  }
+
+  async register(userRegistrationInput) {
+    const email = userRegistrationInput.email;
+    UserEntity.validateEmail(email);
+    const storedUser = await this.userService.getToLogin(email);
+
+    if (storedUser) {
+      throw new RegisteredEmailError(
+        `The ${email} email has already been registered`,
+      );
     }
 
-    async register(userRegistrationInput) {
-        const email = userRegistrationInput.email
-        UserEntity.validateEmail(email)
-        const storedUser = await this.userService.getToLogin(email)
+    const {salt, hash} = this.encryptionService.encrypt(
+      userRegistrationInput.password,
+    );
+    const nameValueObject = new NameValueObject(
+      userRegistrationInput.firstName,
+      userRegistrationInput.lastName,
+    );
+    const contactValueObject = new ContactValueObject(
+      userRegistrationInput.ddd,
+      userRegistrationInput.number,
+    );
+    const user = new UserEntity(
+      email,
+      salt,
+      hash,
+      nameValueObject,
+      contactValueObject,
+    );
 
-        if (storedUser) {
-            throw new RegisteredEmailError(`The ${email} email has already been registered`)
-        }
-
-        const { salt, hash } = this.encryptionService.encrypt(userRegistrationInput.password)
-        const nameValueObject = new NameValueObject(
-            userRegistrationInput.firstName,
-            userRegistrationInput.lastName
-        )
-        const contactValueObject = new ContactValueObject(
-            userRegistrationInput.ddd,
-            userRegistrationInput.number
-        )
-        const user = new UserEntity(
-            email,
-            salt,
-            hash,
-            nameValueObject,
-            contactValueObject
-        )
-
-        await this.userService.add(user)
-    }
-}
-
-module.exports = {
-    RegistrationService
+    await this.userService.add(user);
+  }
 }
