@@ -1,6 +1,3 @@
-module.exports = function(api) {
-  api.cache(true);
-  return {
-    presets: ['babel-preset-expo'],
-  };
+module.exports = {
+  presets: ['module:metro-react-native-babel-preset'],
 };
