@@ -1,13 +1,9 @@
-class LocalizationInput {
-    state
-    city
+export default class LocalizationInput {
+  state;
+  city;
 
-    constructor(state, city) {
-        this.state = state
-        this.city = city
-    }
-}
-
-module.exports = {
-    LocalizationInput
+  constructor(state, city) {
+    this.state = state;
+    this.city = city;
+  }
 }
