@@ -1,13 +1,9 @@
-class UserLoginInput {
-    email
-    password
+export default class UserLoginInput {
+  email;
+  password;
 
-    constructor(email, password) {
-        this.email = email
-        this.password = password
-    }
-}
-
-module.exports = {
-    UserLoginInput
+  constructor(email, password) {
+    this.email = email;
+    this.password = password;
+  }
 }
