@@ -1,28 +1,17 @@
-class UserRegistrationInput {
-    email
-    password
-    firstName
-    lastName
-    ddd
-    number
+export default class UserRegistrationInput {
+  email;
+  password;
+  firstName;
+  lastName;
+  ddd;
+  number;
 
-    constructor(
-        email,
-        password,
-        firstName,
-        lastName,
-        ddd,
-        number
-    ) {
-        this.email = email
-        this.password = password
-        this.firstName = firstName
-        this.lastName = lastName
-        this.ddd = ddd
-        this.number = number
-    }
-}
-
-module.exports = {
-    UserRegistrationInput
+  constructor(email, password, firstName, lastName, ddd, number) {
+    this.email = email;
+    this.password = password;
+    this.firstName = firstName;
+    this.lastName = lastName;
+    this.ddd = ddd;
+    this.number = number;
+  }
 }
