@@ -1,19 +1,15 @@
-class UserService {
-    userRepository
+export default class UserService {
+  userRepository;
 
-    constructor(userRepository) {
-        this.userRepository = userRepository
-    }
+  constructor(userRepository) {
+    this.userRepository = userRepository;
+  }
 
-    async add(userEntity) {
-        await this.userRepository.add(userEntity)
-    }
+  async add(userEntity) {
+    await this.userRepository.add(userEntity);
+  }
 
-    getToLogin(email) {
-        return this.userRepository.getByEmail(email)
-    }
-}
-
-module.exports = {
-    UserService
+  getToLogin(email) {
+    return this.userRepository.getByEmail(email);
+  }
 }
