@@ -1,40 +1,37 @@
-class SetupDatabase {
-    context
+export default class SetupDatabase {
+  context;
 
-    constructor(context) {
-        this.context = context
-    }
+  constructor(context) {
+    this.context = context;
+  }
 
-    async setup() {
-        const database = await this.context.createDatabase()
-        database.serialize(() => {
-            this.createUserTable(database)
-            this.createAdoptionRegistrationTable(database)
-        })
-        return this.context.closeDatabase(database)
-    }
+  async setup() {
+    const database = await this.context.createDatabase();
+    database.serialize(() => {
+      this.createUserTable(database);
+      this.createAdoptionRegistrationTable(database);
+    });
+    return this.context.closeDatabase(database);
+  }
 
-    createUserTable(database) {
-        const sql = `
+  async createUserTable() {
+    const sql = `
         CREATE TABLE IF NOT EXISTS [User] (
             Id INTEGER PRIMARY KEY AUTOINCREMENT,
             Email TEXT,
-            SaltHash BLOB,
-            StoredHash BLOB,
+            SaltHash TEXT,
+            StoredHash TEXT,
             Name_FirstName TEXT,
             Name_LastName TEXT,
             Contact_Ddd TEXT,
             Contact_Number TEXT
-        )`
-        database.run(sql, [], (error) => {
-            if (error) {
-                console.log(error.message)
-            }
-        })
-    }
+        )`;
+    const database = await this.context.createDatabase();
+    return this.context.run(database, sql, []);
+  }
 
-    createAdoptionRegistrationTable(database) {
-        const sql = `
+  async createAdoptionRegistrationTable() {
+    const sql = `
         CREATE TABLE IF NOT EXISTS [AdoptionRegistration] (
             Id INTEGER PRIMARY KEY AUTOINCREMENT,
             Responsible_Id INTEGER,
@@ -44,15 +41,8 @@ class SetupDatabase {
             Animal_Race TEXT,
             Animal_Size Text,
             FOREIGN KEY (Responsible_Id) REFERENCES User (Id)
-        )`
-        database.run(sql, [], (error) => {
-            if (error) {
-                console.log(error.message)
-            }
-        })
-    }
-}
-
-module.exports = {
-    SetupDatabase
+        )`;
+    const database = await this.context.createDatabase();
+    return this.context.run(database, sql, []);
+  }
 }
